Moved enquiry components out of housing and into projectCommon (cross initiative) folder
