+++ frontend/src/components/projectCommon/enquiry/EnquiryCard.vue
